Add support for SQLite and/or DuckDB output
- Added dbtypes config option to select database type(s)
- Updated schema to use BIGINT for IP addresses (required for DuckDB)
- Refactored main.go to support multiple database outputs simultaneously
- Added DuckDB driver dependency
- Updated README with database configuration instructions

+++ pkg/model/cidr.go
@@ -17,8 +17,8 @@ type Cidr struct {
 const createNetDB string = `
 CREATE TABLE IF NOT EXISTS net (
 	net TEXT PRIMARY KEY,
-	start_ip INT NOT NULL,
-	end_ip INT NOT NULL,
+	start_ip BIGINT NOT NULL,
+	end_ip BIGINT NOT NULL,
 	url TEXT NOT NULL,
 	cloudplatform TEXT NOT NULL,
 	iptype TEXT NOT NULL
